split manifold and boolean tests
